cleaner job should not report error when there is nothing to clean

+++ consumer-authorization/src/main/java/eu/arrowhead/authorization/quartz/CleanerJob.java
@@ -121,7 +121,9 @@ public class CleanerJob implements Job {
 				hasMore = currentPage < pageOfHeaders.getTotalPages() - 1;
 			} while (hasMore);
 
-			tokenHeaderDbService.deleteById(toDelete);
+			if (!toDelete.isEmpty()) {
+				tokenHeaderDbService.deleteById(toDelete);
+			}
 		} catch (final Exception ex) {
 			logger.debug(ex);
 			logger.error("Cleaner job error: " + ex.getMessage());
